# strict/swagger.json
{
    "openapi": "3.0.3",
    "paths": {
...
        "/reparation/option/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"[{\\\"$match\\\":{}}]\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationController"
                ]
            }
        },
        "/reparation/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationController"
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "properties": {
                                "_bsontype": {
                                    "type": "object",
                                    "example": "\"ObjectID\""
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationController"
                ]
            }
        },
        "/reparation": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "voitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "startDate": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "endDate": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "avancement": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "cost": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "properties": {}
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "voitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "startDate": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "endDate": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "avancement": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "cost": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "properties": {}
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reparationDetail/option/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"[{\\\"$match\\\":{}}]\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationDetailController"
                ]
            }
        },
        "/reparationDetail": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationDetailController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reparationId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "marquePieceId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "quantity": {
                                        "type": "number"
                                    },
                                    "etat": {
                                        "type": "number"
                                    },
                                    "cost": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationDetailController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reparationId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "marquePieceId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "quantity": {
                                        "type": "number"
                                    },
                                    "etat": {
                                        "type": "number"
                                    },
                                    "cost": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reparationDetail/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationDetailController"
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "properties": {
                                "_bsontype": {
                                    "type": "object",
                                    "example": "\"ObjectID\""
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ReparationDetailController"
                ]
            }
        },
        "/responsable/option/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"[{\\\"$match\\\":{}}]\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ResponsableController"
                ]
            }
        },
        "/responsable/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ResponsableController"
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "properties": {
                                "_bsontype": {
                                    "type": "object",
                                    "example": "\"ObjectID\""
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ResponsableController"
                ]
            }
        },
        "/responsable": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ResponsableController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "properties": {}
                                    },
                                    "mdp": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "ResponsableController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "properties": {}
                                    },
                                    "mdp": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}